+++ packages/handlersjs-http/lib/servers/node/node-http-request-response.handler.ts
@@ -1,10 +1,12 @@
 import { getLoggerFor } from '@digita-ai/handlersjs-logging';
 import { Observable, of, Subject, throwError } from 'rxjs';
 import { map, switchMap, toArray, catchError } from 'rxjs/operators';
+import { BadRequestHttpError } from '../../errors/bad-request-http-error';
 import { HttpHandler } from '../../models/http-handler';
 import { HttpHandlerContext } from '../../models/http-handler-context';
 import { HttpHandlerRequest } from '../../models/http-handler-request';
 import { HttpMethods } from '../../models/http-method';
+import { statusCodes } from '../../handlers/error.handler';
 import { NodeHttpStreamsHandler } from './node-http-streams.handler';
 import { NodeHttpStreams } from './node-http-streams.model';
 
@@ -21,7 +23,11 @@ export class NodeHttpRequestResponseHandler implements NodeHttpStreamsHandler {
    *
    * @param { HttpHandler } httpHandler - the handler through which to pass incoming requests.
    */
-  constructor(private httpHandler: HttpHandler) {
+  constructor(
+    private httpHandler: HttpHandler,
+    private poweredBy = 'handlers.js',
+    private hsts?: { maxAge: number; includeSubDomains: boolean },
+  ) {
 
     if (!httpHandler) {
 
@@ -37,9 +43,33 @@ export class NodeHttpRequestResponseHandler implements NodeHttpStreamsHandler {
     // case 'application/':
     //   return JSON.parse(`{"${decodeURIComponent(body).replace(/"/g, '\\"').replace(/&/g, '","').replace(/=/g, '":"')}"}`);
 
+    this.logger.info('Parsing request body', { body, contentType });
+
     if (contentType?.startsWith('application/json')) {
 
-      return JSON.parse(body);
+      try {
+
+        return JSON.parse(body);
+
+      } catch(error: any) {
+
+        throw new BadRequestHttpError(error.message);
+
+      }
+
+    }
+
+    return body;
+
+  }
+
+  private parseResponseBody(body: unknown, contentType?: string) {
+
+    this.logger.info('Parsing response body', { body, contentType });
+
+    if (contentType?.startsWith('application/json')) {
+
+      return typeof body === 'string' || body instanceof Buffer ? body : JSON.stringify(body);
 
     } else {
 
@@ -142,9 +172,12 @@ export class NodeHttpRequestResponseHandler implements NodeHttpStreamsHandler {
       }),
       catchError((error) => {
 
-        this.logger.debug('Internal server error: ', error);
+        const status = error?.statusCode ?? error.status;
+        const message = error?.message ?? error.body;
 
-        return of({ headers: {}, ...error, body: 'Internal Server Error', status: 500 });
+        this.logger.debug(`${error.name}:`, error);
+
+        return of({ headers: {}, ... error, body: message ?? 'Internal Server Error', status: statusCodes[status] ? status : 500 });
 
       }),
       switchMap((response) => {
@@ -183,6 +216,8 @@ export class NodeHttpRequestResponseHandler implements NodeHttpStreamsHandler {
         const extraHeaders = {
           ... (body !== undefined && body !== null && !response.headers['content-type'] && !response.headers['Content-Type'] && typeof response.body !== 'string' && !(response.body instanceof Buffer)) && { 'content-type': 'application/json' },
           ... (body !== undefined && body !== null) && { 'content-length': Buffer.byteLength(body, charsetString).toString() },
+          ... (this.hsts) && { 'strict-transport-security': `max-age=${this.hsts.maxAge}${this.hsts.includeSubDomains ? '; includeSubDomains' : ''}` },
+          'x-powered-by': this.poweredBy,
         };
 
         return of({
@@ -201,7 +236,10 @@ export class NodeHttpRequestResponseHandler implements NodeHttpStreamsHandler {
 
         if (response.body !== undefined && response.body !== null) {
 
-          nodeHttpStreams.responseStream.write(response.body);
+          const contentTypeHeader = response.headers['content-type'] || response.headers['Content-Type'];
+
+          const body = this.parseResponseBody(response.body, contentTypeHeader);
+          nodeHttpStreams.responseStream.write(body);
 
         }